Cite boatright-ruiz paper in README Remodel tables files to expand testing
